Separate out XML processor
